+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Added optional Visual Crossing weather service.
 * Removed soon-to-be-discontinued Dark Sky weather service.
 * Added ability to use a pool of multiple NTP servers.
+* Added client option for choosing from among available weather services (if API keys have been provided to make more than one service available).
 
 ## 2.11.0